+++ packages/zenrouter_file_generator/example/lib/routes/_coordinator.dart
@@ -1,4 +1,9 @@
 import 'package:zenrouter_file_annotation/zenrouter_file_annotation.dart';
 
-@ZenCoordinator(name: 'AppCoordinator', routeBase: 'AppRoute')
+@ZenCoordinator(
+  name: 'AppCoordinator',
+  routeBase: 'AppRoute',
+  routeBasePath: '_route.dart',
+  deferredImport: true,
+)
 class CoordinatorConfig {}